Fixed the 1eNg and 2eNg topologies

+++ source/falaise/snemo/reconstruction/topology_1eNg_builder.cc
@@ -42,10 +42,7 @@ namespace snemo {
 
       dynamic_cast<snemo::datamodel::topology_1eNg_pattern &>(pattern_).set_number_of_gammas(ngammas);
 
-      // Either this way
-      // for (size_t i_gamma = 1; i_gamma <= h_pattern_->get_number_of_gammas();++i_gamma) {
-      // Or this way but less evolutive
-      for (size_t i_gamma = 1; i_gamma <= ngammas;++i_gamma) {
+      for (int i_gamma = 1; i_gamma <= ngammas;++i_gamma) {
         std::ostringstream oss;
         oss << "g" << i_gamma;
